+++ ManaKit/Classes/CardTableViewCell.swift
@@ -20,7 +20,8 @@ public class CardTableViewCell: UITableViewCell {
             updateDataDisplay()
         }
     }
-    
+    public var faceOrder = 0
+
     // MARK: Outlets
     @IBOutlet weak var thumbnailImage: UIImageView!
     @IBOutlet weak var nameLabel: UILabel!
@@ -124,16 +125,20 @@ public class CardTableViewCell: UITableViewCell {
         // thumbnail image
         if let croppedImage = ManaKit.sharedInstance.cardImage(card,
                                                                imageType: .artCrop,
+                                                               faceOrder: faceOrder,
                                                                roundCornered: false) {
             thumbnailImage.image = croppedImage
         } else {
             thumbnailImage.image = ManaKit.sharedInstance.imageFromFramework(imageName: .cardBackCropped)
 
             firstly {
-                ManaKit.sharedInstance.downloadImage(ofCard: card, imageType: .artCrop)
+                ManaKit.sharedInstance.downloadImage(ofCard: card,
+                                                     imageType: .artCrop,
+                                                     faceOrder: faceOrder)
             }.done {
                 guard let image = ManaKit.sharedInstance.cardImage(card,
                                                                    imageType: .artCrop,
+                                                                   faceOrder: self.faceOrder,
                                                                    roundCornered: false) else {
                     return
                 }